+++ pulse-binding-mainloop-glib/CHANGELOG.md
@@ -1,3 +1,7 @@
+# <unreleased>
+
+ * Changed `Mainloop::new()` to take a ref of `glib::MainContext` not `glib_sys::GMainContext`
+
 # 2.8.0 (August 13th, 2019)
 
  * Addressed long standing todo item of linking `GMainContext` to an actual glib crate.